services/tracking_service/requirements.txt
fastapi==0.115.5
uvicorn==0.30.1
pyjwt==2.9.0
cryptography==42.0.8
